requirements.txt
openai
tiktoken 
chromadb
langchain
python-dotenv
-e .
